Debugger event updates
This moves debugger event reporting more towards a safe api instead of
reporting the raw Win32 structs.

Closely related - the module name (as decided by our debugger) is
reported to callbacks for use in symbol resolution.

The exception callback remains as is for now as it is the most
complicated and heavily used.