Fix potential race while single stepping
When debugging multithreaded targets, there was a potential race in
handling a breakpoint where thread A hits a breakpoint while thread B is
single stepping to the breakpoint.

The fix is, when single stepping, to resume a single thread instead of
all threads that are single stepping.

The step out functionality has also been removed from this layer in part
to keep things simple here. Clients can implement their own step out
logic if needed.