save the EventType enum name for `status top`
A user saw an exception trying to sort EventType in the `status top` UI.
The only place EventType might be sorted is in the Message (table for
rendering event messages).

Since we're only displaying the name of the enum and not doing anything
else, save the name early which removes the issue.