limit the azcopy concurrency
When running tasks with large numbers of machines, on VMs with large numbers of cores, with large numbers of files, we rapidly hit Azure Storage performance caps.

azcopy sets the concurrency rate to 32*NUM_CORES.  On a job with hundreds of VMs, this well exceeds the requests-per-second cap of storage accounts.