Fix some CI flakiness by separating more artifacts (#3250)
Many jobs uploading output into the same artifact causes spurious failures unless the job which consumes the artifact depends upon all of the previous jobs. So, split the uploads into separate artifacts so that the CLI job (which is the main one that fails) can depend upon only a subset of the jobs.

Also add caching to the AFL/AFLPP builds.