Make ServiceConfiguration eagerly evalulated (#3136)
See #3316. This helps diagnosing problems when running the service.

Read all configuration variables in the service upon startup to detect problems as early as possible. This also allows us to remove a bunch of nullability annotations and makes the config easier to use, as well as enforcing stronger types like `Url` and `ResourceIdentifier`.