Fix resolution of sibling .NET DLLs (#3325)
When `LibFuzzerDotNetLoader` loads DLLs that have dependencies, they are not currently resolved correctly, as sibling DLLs in the same directory are not considered. In order to do this, we need to provide a custom `AssemblyLoadContext`. This PR does so, based on the [dotnet core example code](https://learn.microsoft.com/en-us/dotnet/core/tutorials/creating-app-with-plugin-support).

Validated with custom fuzzer implementation; we don't yet have a full test for this in check-pr.

Closes #3326.