Fixup onefuzz-task dependencies (#3552)
Both the bin and lib for `onefuzz-task` include the same modules, which means some code is being compiled twice, and this leads to confusing errors if you update the bin code and not the lib code (e.g. adding new modules or something similar).

Instead, make the bin code depend on the lib code.

Also remove `extern crate` and `macro_use`, which are an older way of doing things. Macros are simply imported now.